requirements.txt
flask
flask-cors
ultralytics
Pillow
numpy
